requirements.txt
asgiref==3.7.2
Django==5.0.8
django-pandas==0.6.7
exceptiongroup==1.2.1
gunicorn==22.0.0
iniconfig==2.0.0
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==2.0.0
packaging==24.0
pandas==2.2.2
pluggy==1.5.0
pytest==8.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
diff-match-patch==20230430
django-import-export==4.1.1
tablib[xlsx]==3.5.0